File.delete do not crashes if files does not exist

+++ src/dstools/pipeline/products/File.py
@@ -45,8 +45,12 @@ class File(Product):
     def delete(self, force=False):
         # force is not used for this product but it is left for API
         # compatibility
-        self.logger.debug(f'Deleting {self._path_to_file}')
-        os.remove(self._path_to_file)
+        if self.exists():
+            self.logger.debug(f'Deleting {self._path_to_file}')
+            os.remove(self._path_to_file)
+        else:
+            self.logger.debug(f'{self._path_to_file} does not exist '
+                              'ignoring...')
 
     @property
     def name(self):
